Ensure NuGet package provider is bootstrapped before use
When WMF5 is freshly installed, the first time `Install-Module` tries
to download a package it will prompt the user to download
`NuGet-anycpu.exe`. The `-Force` parameter does not bypass this prompt which
results in Kitchen appearing to hang during the Verify step.

Running `Get-PackageProvider` will not cause NuGet to be downloaded again
if it is already found on the system.