:bug: Properly handle custom downloads settings
Previously we were assuming the user input was of the form
{ [remotefile] => localfile }
since our default config was of that form.

That format is actually impossible to specify in yaml, so we need to
ensure we're handling both scalar and array keys for our default config
and user input config to work correctly.