(#57) Download Pester zip to modules folder
Prior to this, when Pester can't be installed via PSGet, we download it
directly from Github, and it ends up downloaded to the verifier folder
which causes the Pester module tests to be picked up when we do tests.

Fix is to store and extract the zip in the modules folder instead,
which prevents Invoke-Pester from picking up its own module's tests.