:bug: Workaround test-kitchen download failure
Test Kitchen doesn't `ensure` the file downloads happen, so we can miss
getting the files downloaded, which is not ideal in most cases.

Fix is to use `ensure` if on older `kitchen` versions, and `rescue` the
raised error if on a newer kitchen version to make sure we can get the
files, and we don't get the same files twice if TK fixes the base code.