(#107) Honor Pester ExitCode
Pester always sets `$LASTEXITCODE`, so there is no need to manually set
it to the number of failed tests. Further, in the event of a container
failure, the number of failed tests could be 0 when the failure occurs
before enumeration of the tests.

Additionally, this exit code doesn't actually help as test-kitchen does
not display the exit code, and itself exits with 20.